Rename project_core crate to worktree
Make the crate more exclusively about worktree stuff. Separate project
settings from worktree settings, move the project settings back into
the project crate.
