+++ Cargo.toml
@@ -54,7 +54,6 @@ members = [
     "crates/picker",
     "crates/prettier",
     "crates/project",
-    "crates/project_core",
     "crates/project_panel",
     "crates/project_symbols",
     "crates/quick_action_bar",
@@ -90,6 +89,7 @@ members = [
     "crates/vim",
     "crates/welcome",
     "crates/workspace",
+    "crates/worktree",
     "crates/zed",
     "crates/zed_actions",
 
@@ -159,7 +159,7 @@ plugin = { path = "crates/plugin" }
 plugin_macros = { path = "crates/plugin_macros" }
 prettier = { path = "crates/prettier" }
 project = { path = "crates/project" }
-project_core = { path = "crates/project_core" }
+worktree = { path = "crates/worktree" }
 project_panel = { path = "crates/project_panel" }
 project_symbols = { path = "crates/project_symbols" }
 quick_action_bar = { path = "crates/quick_action_bar" }
