+++ crates/collab/src/tests/following_tests.rs
@@ -12,7 +12,7 @@ use gpui::{
 };
 use language::Capability;
 use live_kit_client::MacOSDisplay;
-use project::project_settings::ProjectSettings;
+use project::WorktreeSettings;
 use rpc::proto::PeerId;
 use serde_json::json;
 use settings::SettingsStore;
@@ -1646,8 +1646,8 @@ async fn test_following_into_excluded_file(
     for cx in [&mut cx_a, &mut cx_b] {
         cx.update(|cx| {
             cx.update_global::<SettingsStore, _>(|store, cx| {
-                store.update_user_settings::<ProjectSettings>(cx, |project_settings| {
-                    project_settings.file_scan_exclusions = Some(vec!["**/.git".to_string()]);
+                store.update_user_settings::<WorktreeSettings>(cx, |settings| {
+                    settings.file_scan_exclusions = Some(vec!["**/.git".to_string()]);
                 });
             });
         });
