+++ crates/project/Cargo.toml
@@ -19,7 +19,7 @@ test-support = [
     "settings/test-support",
     "text/test-support",
     "prettier/test-support",
-    "project_core/test-support",
+    "worktree/test-support",
     "gpui/test-support",
 ]
 
@@ -44,10 +44,11 @@ node_runtime.workspace = true
 parking_lot.workspace = true
 postage.workspace = true
 prettier.workspace = true
-project_core.workspace = true
+worktree.workspace = true
 rand.workspace = true
 regex.workspace = true
 rpc.workspace = true
+schemars.workspace = true
 task.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -72,7 +73,7 @@ release_channel.workspace = true
 lsp = { workspace = true, features = ["test-support"] }
 prettier = { workspace = true, features = ["test-support"] }
 pretty_assertions.workspace = true
-project_core = { workspace = true, features = ["test-support"] }
+worktree = { workspace = true, features = ["test-support"] }
 rpc = { workspace = true, features = ["test-support"] }
 settings = { workspace = true, features = ["test-support"] }
 unindent.workspace = true
