+++ crates/worktree/src/ignore.rs
